chore: complete migration to new intent label
