+++ botfront/imports/ui/components/nlu/activity/Activity.jsx
@@ -3,7 +3,7 @@ import PropTypes from 'prop-types';
 import { connect } from 'react-redux';
 import moment from 'moment';
 import { Message, Segment } from 'semantic-ui-react';
-import IntentLabel from '../common/NewIntentLabel';
+import IntentLabel from '../common/IntentLabel';
 import UserUtteranceViewer from '../common/UserUtteranceViewer';
 import { useActivity, useDeleteActivity, useUpsertActivity } from './hooks';
